WIP
- Moved subscription-related methods to base subscription class
- implemented support for legacy payment method ID requests that confirm payment intent immediately
- Added support for checkout and stripe elements with support for themes and
- Removed dependency on jQuery in payment form
-
