+++ CHANGELOG-WIP.md
@@ -5,3 +5,7 @@
 - Added the Checkout gateway for redirect based Checkout payments.
 - Added the `stripe/billing-portal` action which will redirect a logged-in user to their stripe billing portal page.
 - We now update the stripe customer when a user is updated.
+
+- Removed `craft\commerce\stripe\events\BuildGatewayRequestEvent::$metadata`. Use `BuildGatewayRequestEvent::$request`
+  instead.
+- Removed `\craft\commerce\stripe\base\Gateway::normalizePaymentToken`.
