+++ src/gateways/PaymentIntents.php
@@ -11,7 +11,6 @@ use Craft;
 use craft\commerce\base\Plan as BasePlan;
 use craft\commerce\base\RequestResponseInterface;
 use craft\commerce\base\SubscriptionResponseInterface;
-use craft\commerce\behaviors\CustomerBehavior;
 use craft\commerce\elements\Subscription;
 use craft\commerce\errors\SubscriptionException;
 use craft\commerce\models\payments\BasePaymentForm;
@@ -21,16 +20,18 @@ use craft\commerce\models\Transaction;
 use craft\commerce\Plugin as Commerce;
 use craft\commerce\stripe\base\SubscriptionGateway as BaseGateway;
 use craft\commerce\stripe\errors\PaymentSourceException;
+use craft\commerce\stripe\events\BuildGatewayRequestEvent;
 use craft\commerce\stripe\events\PaymentIntentConfirmationEvent;
+use craft\commerce\stripe\events\PaymentIntentRequestEvent;
 use craft\commerce\stripe\events\SubscriptionRequestEvent;
-use craft\commerce\stripe\models\forms\payment\PaymentIntent as PaymentForm;
+use craft\commerce\stripe\models\forms\payment\PaymentIntent as PaymentIntentForm;
 use craft\commerce\stripe\models\forms\Subscription as SubscriptionForm;
-use craft\commerce\stripe\models\PaymentIntent as PaymentIntentModel;
 use craft\commerce\stripe\Plugin as StripePlugin;
+use craft\commerce\stripe\responses\CheckoutSessionResponse;
 use craft\commerce\stripe\responses\PaymentIntentResponse;
+use craft\commerce\stripe\web\assets\elementsform\ElementsFormAsset;
 use craft\commerce\stripe\web\assets\intentsform\IntentsFormAsset;
 use craft\elements\User;
-use craft\errors\ElementNotFoundException;
 use craft\helpers\Json;
 use craft\helpers\StringHelper;
 use craft\helpers\UrlHelper;
@@ -55,17 +56,53 @@ use function count;
  **/
 class PaymentIntents extends BaseGateway
 {
+
+    public const PAYMENT_FORM_TYPE_CARD = 'card'; // default
+    public const PAYMENT_FORM_TYPE_CHECKOUT = 'checkout';
+    public const PAYMENT_FORM_TYPE_ELEMENTS = 'elements';
+
+    /**
+     * @var string
+     */
+    public string $_formType = 'card';
+
     /**
      * @event BeforeConfirmPaymentIntent The event that is triggered before a PaymentIntent is confirmed
      */
     public const EVENT_BEFORE_CONFIRM_PAYMENT_INTENT = 'beforeConfirmPaymentIntent';
+
 
     /**
      * @inheritdoc
      */
     public static function displayName(): string
     {
-        return Craft::t('commerce-stripe', 'Stripe (Legacy Card)');
+        return Craft::t('commerce-stripe', 'Stripe');
+    }
+
+    /**
+     * @return string
+     */
+    public function getFormType(): string
+    {
+        return $this->_formType;
+    }
+
+    /**
+     * @param string $formType
+     * @return void
+     */
+    public function setFormType(string $formType): void
+    {
+        $this->_formType = $formType;
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function showPaymentFormSubmitButton(): bool
+    {
+        return $this->getFormType() == self::PAYMENT_FORM_TYPE_CHECKOUT;
     }
 
     /**
@@ -75,41 +112,37 @@ class PaymentIntents extends BaseGateway
     {
         $defaults = [
             'gateway' => $this,
-            'paymentForm' => $this->getPaymentFormModel(),
-            'scenario' => 'payment',
             'handle' => $this->handle,
+            'appearance' => [
+                'theme' => 'stripe'
+            ],
+            'layout' => [
+                'type' => 'tabs',
+            ],
+            'submitButtonText' => Craft::t('commerce', 'Pay'),
+            'submitButtonClasses' => '',
+            'errorMessageClasses' => '',
+            'paymentFormType' => $this->getFormType(),
+            'paymentMethodTypes' => ['card'],
         ];
 
         $params = array_merge($defaults, $params);
 
-        // If there's no order passed, add the current cart if we're not messing around in backend.
-        if (!isset($params['order']) && !Craft::$app->getRequest()->getIsCpRequest()) {
-            if ($cart = Commerce::getInstance()->getCarts()->getCart()) {
-                $billingAddress = $cart->getBillingAddress();
-
-                /** @var User|CustomerBehavior|null $user */
-                $user = $cart->getCustomer();
-                if (!$billingAddress && $user) {
-                    $billingAddress = $user->getPrimaryBillingAddress();
-                }
-            }
-        } else {
-            $billingAddress = $params['order']->getBillingAddress();
-        }
-
-        if ($billingAddress) {
-            $params['billingAddress'] = $billingAddress;
-        }
+        // Set it in memory on the gateway for this request
+        $this->setFormType($params['paymentFormType']);
 
         $view = Craft::$app->getView();
-
         $previousMode = $view->getTemplateMode();
         $view->setTemplateMode(View::TEMPLATE_MODE_CP);
 
         $view->registerScript('', View::POS_END, ['src' => 'https://js.stripe.com/v3/']); // we need this to load at end of body
-        $view->registerAssetBundle(IntentsFormAsset::class);
 
-        $html = $view->renderTemplate('commerce-stripe/paymentForms/intentsForm', $params);
+        if ($this->getFormType() == self::PAYMENT_FORM_TYPE_CHECKOUT) {
+            $html = $view->renderTemplate('commerce-stripe/paymentForms/checkoutForm', $params);;
+        } else {
+            $view->registerAssetBundle(ElementsFormAsset::class);
+            $html = $view->renderTemplate('commerce-stripe/paymentForms/elementsForm', $params);
+        }
 
         $view->setTemplateMode($previousMode);
 
@@ -133,7 +166,7 @@ class PaymentIntents extends BaseGateway
      */
     public function getPaymentFormModel(): BasePaymentForm
     {
-        return new PaymentForm();
+        return new PaymentIntentForm();
     }
 
     /**
@@ -149,31 +182,18 @@ class PaymentIntents extends BaseGateway
      */
     public function completePurchase(Transaction $transaction): RequestResponseInterface
     {
-        $paymentIntentReference = Craft::$app->getRequest()->getParam('payment_intent');
-        $stripePaymentIntent = $this->getStripeClient()->paymentIntents->retrieve($paymentIntentReference);
+        $data = Json::decodeIfJson($transaction->response);
 
-        // Update the intent with the latest.
-        $paymentIntentsService = StripePlugin::getInstance()->getPaymentIntents();
-
-        $paymentIntent = $paymentIntentsService->getPaymentIntentByReference($paymentIntentReference);
-
-        // Make sure we have the payment intent before we attempt to do anything with it.
-        if ($paymentIntent) {
-            $paymentIntent->intentData = Json::encode($stripePaymentIntent->toArray());
-            $paymentIntentsService->savePaymentIntent($paymentIntent);
+        if ($data['object'] == 'payment_intent') {
+            $paymentIntent = $this->getStripeClient()->paymentIntents->retrieve($data['id']);
+        } else {
+            // Likely a checkout object
+            $checkoutSession = $this->getStripeClient()->checkout->sessions->retrieve($data['id']);
+            $paymentIntent = $checkoutSession['payment_intent'];
+            $paymentIntent = $this->getStripeClient()->paymentIntents->retrieve($paymentIntent);
         }
 
-        $intentData = $stripePaymentIntent->toArray();
-
-        if (!empty($intentData['payment_method'])) {
-            try {
-                $this->_confirmPaymentIntent($stripePaymentIntent, $transaction);
-            } catch (Exception $exception) {
-                return $this->createPaymentResponseFromError($exception);
-            }
-        }
-
-        return $this->createPaymentResponseFromApiResource($stripePaymentIntent);
+        return $this->createPaymentResponseFromApiResource($paymentIntent);
     }
 
     /**
@@ -211,7 +231,6 @@ class PaymentIntents extends BaseGateway
         }
 
         $stripePaymentIntent = $this->getStripeClient()->paymentIntents->retrieve($transaction->reference);
-        $paymentIntentsService = StripePlugin::getInstance()->getPaymentIntents();
 
         try {
             if ($stripePaymentIntent->status == 'succeeded') {
@@ -329,25 +348,6 @@ class PaymentIntents extends BaseGateway
         return true;
     }
 
-    /**
-     * @inheritdoc
-     */
-    public function getBillingIssueDescription(Subscription $subscription): string
-    {
-        $subscriptionData = $this->_getExpandedSubscriptionData($subscription);
-        $intentData = $subscriptionData['latest_invoice']['payment_intent'];
-
-        if (in_array($subscriptionData['status'], ['incomplete', 'past_due', 'unpaid'])) {
-            switch ($intentData['status']) {
-                case 'requires_payment_method':
-                    return $subscription->hasStarted ? Craft::t('commerce-stripe', 'To resume the subscription, please provide a valid payment method.') : Craft::t('commerce-stripe', 'To start the subscription, please provide a valid payment method.');
-                case 'requires_action':
-                    return $subscription->hasStarted ? Craft::t('commerce-stripe', 'To resume the subscription, please complete 3DS authentication.') : Craft::t('commerce-stripe', 'To start the subscription, please complete 3DS authentication.');
-            }
-        }
-
-        return '';
-    }
 
     /**
      * @inheritdoc
@@ -374,165 +374,181 @@ class PaymentIntents extends BaseGateway
     /**
      * @inheritdoc
      */
-    public function getHasBillingIssues(Subscription $subscription): bool
-    {
-        $subscription = $this->refreshSubscriptionData($subscription);
-        $subscriptionData = $subscription->getSubscriptionData();
-        $intentData = $subscriptionData['latest_invoice']['payment_intent'];
-
-        return in_array($subscriptionData['status'], ['incomplete', 'past_due', 'unpaid']) && in_array($intentData['status'], ['requires_payment_method', 'requires_confirmation', 'requires_action']);
-    }
-
-    /**
-     * @inheritdoc
-     */
     public function handleWebhook(array $data): void
     {
-        switch ($data['type']) {
-            case 'invoice.payment_failed':
-                $this->handleInvoiceFailed($data);
-                break;
-        }
+//        switch ($data['type']) {
+//            case 'invoice.payment_failed':
+//                $this->handleInvoiceFailed($data);
+//                break;
+//        }
 
         parent::handleWebhook($data);
     }
 
     /**
-     * Handle a failed invoice by updating the subscription data for the subscription it failed.
-     *
-     * @param array $data
-     * @throws Throwable
-     * @throws ElementNotFoundException
-     * @throws \yii\base\Exception
-     */
-    protected function handleInvoiceFailed(array $data): void
-    {
-        $stripeInvoice = $data['data']['object'];
-
-        // Sanity check
-        if ($stripeInvoice['paid']) {
-            return;
-        }
-
-        $subscriptionReference = $stripeInvoice['subscription'] ?? null;
-
-        if (!$subscriptionReference || !($subscription = Subscription::find()->status(null)->reference($subscriptionReference)->one())) {
-            Craft::warning('Subscription with the reference “' . $subscriptionReference . '” not found when processing webhook ' . $data['id'], 'stripe');
-
-            return;
-        }
-
-        $this->refreshSubscriptionData($subscription);
-    }
-
-    /**
-     * @inheritdoc
-     */
-    protected function handleSubscriptionUpdated(array $data): void
-    {
-        // Fetch expanded data
-        $stripeSubscription = $this->getStripeClient()->subscriptions->retrieve(
-            $data['data']['object']['id'],
-            [
-                'expand' => ['latest_invoice.payment_intent'],
-            ]);
-
-        // And nonchalantly replace it, before calling parent.
-        $data['data']['object'] = $stripeSubscription->toArray();
-
-        parent::handleSubscriptionUpdated($data);
-    }
-
-    /**
-     * @inheritdoc
-     */
-    protected function authorizeOrPurchase(Transaction $transaction, BasePaymentForm $form, bool $capture = true): RequestResponseInterface
-    {
-        /** @var PaymentForm $form */
-        $requestData = $this->buildRequestData($transaction);
-        $paymentMethodId = $form->paymentMethodId;
-
-        $customer = null;
-        $paymentIntent = null;
-
-        $stripePlugin = StripePlugin::getInstance();
-
-        if ($form->customer) {
-            $requestData['customer'] = $form->customer;
-            $customer = $stripePlugin->getCustomers()->getCustomerByReference($form->customer);
-        } elseif ($user = $transaction->getOrder()->getCustomer()) {
-            $customer = $stripePlugin->getCustomers()->getCustomer($this->id, $user);
-            $requestData['customer'] = $customer->reference;
-        }
-
-        $requestData['payment_method'] = $paymentMethodId;
-
-        $paymentIntentService = $stripePlugin->getPaymentIntents();
-
-        try {
-            // If this is a customer that's logged in, attempt to continue the timeline
-            if ($customer) {
-                $paymentIntent = $paymentIntentService->getPaymentIntent($this->id, $transaction->orderId, $customer->id, $transaction->hash);
-            }
-
-            // If a payment intent exists, update that.
-            if ($paymentIntent) {
-                $stripePaymentIntent = $this->getStripeClient()->paymentIntents->update($paymentIntent->reference, $requestData, ['idempotency_key' => $transaction->hash]);
-            } else {
-                $requestData['capture_method'] = $capture ? 'automatic' : 'manual';
-                $requestData['confirmation_method'] = 'manual';
-                $requestData['confirm'] = false;
-
-                $stripePaymentIntent = $this->getStripeClient()->paymentIntents->create($requestData, ['idempotency_key' => $transaction->hash]);
-
-                if ($customer) {
-                    $paymentIntent = new PaymentIntentModel([
-                        'orderId' => $transaction->orderId,
-                        'transactionHash' => $transaction->hash,
-                        'customerId' => $customer->id,
-                        'gatewayId' => $this->id,
-                        'reference' => $stripePaymentIntent->id,
-                    ]);
-                }
-            }
-
-            if ($paymentIntent) {
-                // Save data before confirming.
-                $paymentIntent->intentData = Json::encode($stripePaymentIntent->toArray());
-                $paymentIntentService->savePaymentIntent($paymentIntent);
-            }
-
-            $this->_confirmPaymentIntent($stripePaymentIntent, $transaction);
-
-            return $this->createPaymentResponseFromApiResource($stripePaymentIntent);
-        } catch (Exception $exception) {
-            return $this->createPaymentResponseFromError($exception);
-        }
-    }
-
-    /**
-     * Refresh a subscription's data.
-     *
-     * @param Subscription $subscription
-     * @return Subscription
+     * @param Transaction $transaction
+     * @param int $amount
+     * @param bool $capture
+     * @param array $metadata
+     * @param ?User $user
+     * @return CheckoutSessionResponse
      * @throws \Stripe\Exception\ApiErrorException
-     * @throws \Throwable
-     * @throws \craft\errors\ElementNotFoundException
-     * @throws \yii\base\Exception
+     * @throws \craft\commerce\stripe\errors\CustomerException
+     * @throws \craft\errors\SiteNotFoundException
+     * @throws \yii\base\InvalidConfigException
      */
-    protected function refreshSubscriptionData(Subscription $subscription): Subscription
+    public function handleCheckoutRequest(Transaction $transaction, int $amount, bool $capture, array $metadata, ?User $user): CheckoutSessionResponse
     {
-        $stripeSubscription = $this->getStripeClient()->subscriptions->retrieve(
-            $subscription->reference,
-            [
-                'expand' => ['latest_invoice.payment_intent'],
-            ]);
+        $lineItems = [];
+        $lineItems[] = [
+            'price_data' => [
+                'currency' => $transaction->paymentCurrency,
+                'unit_amount' => $amount,
+                'tax_behavior' => 'unspecified',
+                'product_data' => [
+                    'name' => Craft::$app->getSites()->getCurrentSite()->name . ' Order #' . $transaction->getOrder()->shortNumber,
+                    'metadata' => [
+                        'order_id' => $transaction->getOrder()->id,
+                        'order_number' => $transaction->getOrder()->number,
+                        'order_short_number' => $transaction->getOrder()->shortNumber,
+                    ],
+                ],
+            ],
+            'adjustable_quantity' => [
+                'enabled' => false,
+            ],
+            'quantity' => 1,
+        ];
 
-        $subscription->setSubscriptionData($stripeSubscription->toArray());
-        $this->setSubscriptionStatusData($subscription);
-        Craft::$app->getElements()->saveElement($subscription);
+        $paymentIntentData = [
+            'capture_method' => $capture ? 'automatic' : 'manual',
+        ];
 
-        return $subscription;
+        $data = [
+            'cancel_url' => $transaction->getOrder()->cancelUrl,
+            'success_url' => UrlHelper::actionUrl('commerce/payments/complete-payment', ['commerceTransactionId' => $transaction->id, 'commerceTransactionHash' => $transaction->hash]),
+            'mode' => 'payment',
+            'client_reference_id' => $transaction->hash,
+            'line_items' => $lineItems,
+            'metadata' => $metadata,
+            'allow_promotion_codes' => false,
+            'payment_intent_data' => $paymentIntentData,
+        ];
+
+        $orderCustomer = $transaction->getOrder()->getCustomer();
+        if ($user && $orderCustomer && $user->id == $orderCustomer->id) {
+            $data['customer'] = StripePlugin::getInstance()->getCustomers()->getCustomer($this->id, $transaction->getOrder()->getCustomer())->reference;
+        } else {
+            $data['customer_email'] = $transaction->getOrder()->email;
+        }
+
+        $event = new BuildGatewayRequestEvent([
+            'transaction' => $transaction,
+            'request' => $data,
+        ]);
+
+        if ($this->hasEventHandlers(self::EVENT_BUILD_GATEWAY_REQUEST)) {
+            $this->trigger(self::EVENT_BUILD_GATEWAY_REQUEST, $event);
+        }
+
+        $session = $this->getStripeClient()->checkout->sessions->create($event->request);
+
+        return new CheckoutSessionResponse($session->toArray());
+    }
+
+    /**
+     * @inheritdoc
+     */
+    protected function authorizeOrPurchase(Transaction $transaction, PaymentIntentForm|BasePaymentForm $form, bool $capture = true): RequestResponseInterface
+    {
+        // When we are using the legacy payment flow, we will need to immediately confirm the payment intent later
+        $immediatelyConfirmLegacy = false;
+
+        $currency = Commerce::getInstance()->getCurrencies()->getCurrencyByIso($transaction->paymentCurrency);
+        $user = Craft::$app->getUser()->getIdentity();
+
+        // This is the metadata that will be sent to Stripe for checkout and payment intents
+        $metadata = [
+            'order_id' => $transaction->getOrder()->id,
+            'order_number' => $transaction->getOrder()->number,
+            'order_short_number' => $transaction->getOrder()->shortNumber,
+            'transaction_id' => $transaction->id,
+            'transaction_reference' => $transaction->hash,
+            'description' => Craft::t('commerce-stripe', 'Order') . ' #' . $transaction->orderId,
+        ];
+
+        $appRequest = Craft::$app->getRequest();
+        if (!$appRequest->getIsConsoleRequest()) {
+            $metadata['client_ip'] = $appRequest->getUserIP();
+        }
+
+        // Normalized amount for Stripe into minor units
+        $amount = $transaction->paymentAmount * (10 ** $currency->minorUnit);
+
+        if ($form->paymentFormType == self::PAYMENT_FORM_TYPE_CHECKOUT) {
+            return $this->handleCheckoutRequest($transaction, $amount, $capture, $metadata, $user);
+        }
+
+        // Base information for the payment intent
+        $paymentIntentData = [
+            'amount' => $amount,
+            'currency' => $transaction->paymentCurrency,
+            'confirm' => false,
+            'metadata' => $metadata,
+            'capture_method' => $capture ? 'automatic' : 'manual',
+        ];
+
+        /**
+         * The previous version of the Stripe plugin accepted a payment method ID on initial
+         * payment intent creation. We can attach it to the payment intent
+         * before creating it, and we will immediately confirm it later in this function to match the previous behavior.
+         */
+        if ($form->paymentMethodId) {
+            $paymentIntentData['payment_method'] = $form->paymentMethodId;
+            $immediatelyConfirmLegacy = true;
+        }
+
+        // Which payment methods are allowed on the payment intent
+        if ($form->paymentFormType == self::PAYMENT_FORM_TYPE_ELEMENTS && $form->paymentMethodTypes) {
+            $paymentIntentData['payment_method_types'] = $form->paymentMethodTypes;
+        } elseif ($form->paymentFormType == self::PAYMENT_FORM_TYPE_CARD) {
+            $paymentIntentData['payment_method_types'] = ['card'];
+        } else {
+            $paymentIntentData['automatic_payment_methods'] = ['enabled' => true];
+        }
+
+        // Add the receipt email if enabled
+        if ($this->sendReceiptEmail) {
+            $paymentIntentData['receipt_email'] = $transaction->getOrder()->getEmail();
+        }
+
+        // Add current customer if the user is logged in
+        $orderCustomer = $transaction->getOrder()->getCustomer();
+        if ($user && $orderCustomer && $user->id == $orderCustomer->id) {
+            $paymentIntentData['customer'] = StripePlugin::getInstance()->getCustomers()->getCustomer($this->id, $transaction->getOrder()->getCustomer())->reference;
+        }
+
+        $event = new BuildGatewayRequestEvent([
+            'transaction' => $transaction,
+            'request' => $paymentIntentData,
+        ]);
+
+        if ($this->hasEventHandlers(self::EVENT_BUILD_GATEWAY_REQUEST)) {
+            $this->trigger(self::EVENT_BUILD_GATEWAY_REQUEST, $event);
+
+            // Do not allow these to be modified by event handlers
+            $event->request['amount'] = $paymentIntentData['amount'];
+            $event->request['currency'] = $paymentIntentData['currency'];
+        }
+
+        $paymentIntent = $this->getStripeClient()->paymentIntents->create($event->request);
+
+        if ($immediatelyConfirmLegacy) {
+            // Mutates the payment intent that has been confirmed
+            $this->_confirmPaymentIntent($paymentIntent, $transaction);
+        }
+
+        return new PaymentIntentResponse($paymentIntent->toArray());
     }
 
     /**
@@ -558,26 +574,14 @@ class PaymentIntents extends BaseGateway
     }
 
     /**
-     * Get the expanded subscription data, including payment intent for latest invoice.
-     *
-     * @param Subscription $subscription
-     * @return array
-     * @throws \Stripe\Exception\ApiErrorException
+     * @inheritdoc
      */
-    private function _getExpandedSubscriptionData(Subscription $subscription): array
+    public function supportsPaymentSources(): bool
     {
-        $subscriptionData = $subscription->getSubscriptionData();
-
-        if (empty($subscriptionData['latest_invoice']['payment_intent'])) {
-            $stripeSubscription = $this->getStripeClient()->subscriptions->retrieve(
-                $subscription->reference,
-                [
-                    'expand' => ['latest_invoice.payment_intent'],
-                ]);
-            $subscriptionData = $stripeSubscription->toArray();
-            $subscription->setSubscriptionData($subscriptionData);
+        if ($this->_formType == self::PAYMENT_FORM_TYPE_CARD) {
+            return true;
         }
 
-        return $subscriptionData;
+        return parent::supportsPaymentSources();
     }
 }
