+++ src/templates/gatewaySettings/intentsSettings.twig
@@ -1,4 +1,4 @@
-{% from "_includes/forms" import autosuggestField, booleanMenuField %}
+{% from "_includes/forms" import autosuggestField, booleanMenuField, selectField %}
 
 {{ autosuggestField({
     label: "Publishable API Key"|t('commerce-stripe'),
