+++ src/templates/paymentForms/elementsForm.twig
@@ -1,16 +1,19 @@
-<div id="payment-form"
-     class="stripe-payment-elements-form hidden"
+<div class="stripe-payment-elements-form hidden"
      data-publishablekey="{{ parseEnv(gateway.getPublishableKey()) }}"
      data-complete-action-url="{{ actionUrl('commerce/payments/complete-payment') }}"
      data-payment-form-namespace="{{ handle|commercePaymentFormNamespace }}"
      data-appearance="{{ appearance|json_encode }}"
      data-layout="{{ layout|json_encode }}"
 >
-    <div class="form-fields">
+    {{ hiddenInput('paymentFormType', 'elements') }}
+    <div class="hidden stripe-error-message {{ errorMessageClasses }}">
+        {# Display error message to your customers here #}
+    </div>
+    <div class="stripe-payment-form">
         <div class="stripe-payment-element">
-            <!-- Elements will create form elements here -->
+            {# Elements will create form elements here #}
         </div>
 
-        <button class="stripe-payment-elements-submit-button">Submit</button>
+        <button class="stripe-payment-elements-submit-button {{ submitButtonClasses }}">{{ submitButtonText }}</button>
     </div>
 </div>
