--- src/web/assets/intentsform/css/paymentForm.css
@@ -1,3 +1,0 @@
-.stripe-payment-intents-form {
-    width: 310px;
-}
